chore: address PR review comments for #92
