+++ adws/adwPlanBuild.tsx
@@ -1,20 +1,15 @@
 #!/usr/bin/env npx tsx
 /**
- * ADW Plan & Build - Self-Sufficient Plan+Build+PR Orchestrator
+ * ADW Plan & Build - Plan+Build+PR Orchestrator
  *
  * Usage: npx tsx adws/adwPlanBuild.tsx <github-issue-number> [adw-id]
  *
  * Workflow:
- * 1. Fetch GitHub issue
- * 2. Setup worktree (with latest code from default branch)
- * 3. Detect recovery state from existing comments
- * 4. Classify issue type (feature, bug, chore, pr_review)
- * 5. Create feature branch: {type}/issue-{number}-{slug}
- * 6. Run Plan Agent: generate implementation plan
- * 7. Commit the plan
- * 8. Run Build Agent: implement the plan
- * 9. Commit the implementation
- * 10. Create Pull Request
+ * 1. Initialize: fetch issue, classify type, setup worktree, initialize state, detect recovery
+ * 2. Plan Phase: classify issue, create branch, run plan agent, commit plan
+ * 3. Build Phase: run build agent, commit implementation
+ * 4. PR Phase: create pull request
+ * 5. Finalize: update state, post completion comment
  *
  * Environment Requirements:
  * - ANTHROPIC_API_KEY: Anthropic API key
@@ -22,36 +17,15 @@
  * - GITHUB_PAT: (Optional) GitHub Personal Access Token
  */
 
+import { generateAdwId } from './core';
 import {
-  log,
-  generateAdwId,
-  ensureLogsDirectory,
-  IssueClassSlashCommand,
-  AgentStateManager,
-  AgentState,
-  setupWorktreeWithLatestCode,
-  handleRecoveryMode,
-  executeClassifyStep,
-  executeCreateBranchStep,
-  executePlanAgentStep,
-  executeCommitPlanStep,
-  readPlanContent,
-  executeBuildAgentStep,
-  executeCommitImplementationStep,
-  executePRCreationStep,
+  initializeWorkflow,
+  executePlanPhase,
+  executeBuildPhase,
+  executePRPhase,
   completeWorkflow,
   handleWorkflowError,
-  WorkflowParams,
-} from './core';
-import {
-  fetchGitHubIssue,
-  WorkflowContext,
-  detectRecoveryState,
-  getDefaultBranch,
-  generateBranchName,
-} from './github';
-import { getPlanFilePath } from './agents';
-import { classifyGitHubIssue } from './triggers/issueClassifier';
+} from './workflowPhases';
 
 /**
  * Prints usage information and exits.
@@ -94,101 +68,15 @@ async function main(): Promise<void> {
   const args = process.argv.slice(2);
   const { issueNumber, adwId } = parseArguments(args);
 
-  log('===================================', 'info');
-  log('ADW Plan & Build Orchestrator', 'info');
-  log(`Issue: #${issueNumber}`, 'info');
-  log(`ADW ID: ${adwId}`, 'info');
-  log('===================================', 'info');
-
-  // Step 1: Fetch issue
-  log('Fetching GitHub issue...', 'info');
-  const issue = await fetchGitHubIssue(issueNumber);
-  log(`Fetched issue: ${issue.title}`, 'success');
-
-  // Step 2: Setup worktree with latest code
-  const defaultBranch = getDefaultBranch();
-  const tempBranchName = generateBranchName(issueNumber, issue.title, '/feature');
-  const worktreePath = setupWorktreeWithLatestCode(tempBranchName, defaultBranch);
-  log(`Default branch: ${defaultBranch}`, 'info');
-
-  // Step 3: Detect recovery state
-  const recoveryState = detectRecoveryState(issue.comments);
-
-  // Step 4: Classify issue
-  log('Classifying issue type...', 'info');
-  const classificationResult = await classifyGitHubIssue(issue);
-  const issueType: IssueClassSlashCommand = classificationResult.issueType;
-  log(`Issue classified as: ${issueType}`, classificationResult.success ? 'success' : 'info');
-
-  // Initialize state
-  const logsDir = ensureLogsDirectory(adwId);
-  const orchestratorStatePath = AgentStateManager.initializeState(adwId, 'plan-build-orchestrator');
-  log(`State: ${orchestratorStatePath}`, 'info');
-  log(`Logs: ${logsDir}`, 'info');
-
-  const initialState: Partial<AgentState> = {
-    adwId,
-    issueNumber,
-    agentName: 'plan-build-orchestrator',
-    execution: AgentStateManager.createExecutionState('running'),
-  };
-  AgentStateManager.writeState(orchestratorStatePath, initialState);
-  AgentStateManager.appendLog(orchestratorStatePath, `Starting ADW Plan & Build workflow for issue #${issueNumber}`);
-
-  // Initialize workflow context
-  const ctx: WorkflowContext = {
-    issueNumber,
-    adwId,
-    issueType,
-  };
-
-  // Build workflow params
-  const params: WorkflowParams = {
-    issueNumber,
-    adwId,
-    issue,
-    issueType,
-    recoveryState,
-    orchestratorStatePath,
-    orchestratorName: 'plan-build-orchestrator',
-    ctx,
-    workingDir: worktreePath,
-    logsDir,
-  };
-
-  // Handle recovery mode
-  handleRecoveryMode(params);
+  const config = await initializeWorkflow(issueNumber, adwId, 'plan-build-orchestrator');
 
   try {
-    let totalCostUsd = 0;
-
-    // === PLAN PHASE ===
-    executeClassifyStep(params);
-    const currentBranch = executeCreateBranchStep(params);
-    totalCostUsd += await executePlanAgentStep(params, currentBranch);
-    executeCommitPlanStep(params);
-
-    // === BUILD PHASE ===
-    const planPath = getPlanFilePath(issueNumber);
-    const planContent = readPlanContent(planPath);
-    totalCostUsd += await executeBuildAgentStep(params, planContent, currentBranch);
-    executeCommitImplementationStep(params);
-
-    // === PR PHASE ===
-    executePRCreationStep(params, defaultBranch);
-
-    // === COMPLETION ===
-    completeWorkflow(orchestratorStatePath, ctx, issueNumber, { totalCostUsd });
-
-    log('===================================', 'info');
-    log('ADW Plan & Build workflow completed!', 'success');
-    if (ctx.prUrl) {
-      log(`PR: ${ctx.prUrl}`, 'info');
-    }
-    log('===================================', 'info');
-
+    const planResult = await executePlanPhase(config);
+    const buildResult = await executeBuildPhase(config);
+    executePRPhase(config);
+    completeWorkflow(config, planResult.costUsd + buildResult.costUsd);
   } catch (error) {
-    handleWorkflowError(error, orchestratorStatePath, ctx, issueNumber, 'Plan & Build workflow');
+    handleWorkflowError(config, error);
   }
 }
